src/daemon/importer/requirements.txt
watchdog==3.0.0
lxml==4.9.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0